--- a/ledger-core/virtual/object/request_table_test.go
+++ b/ledger-core/virtual/object/request_table_test.go
@@ -69,7 +69,7 @@ func TestPendingList(t *testing.T) {
 	RefOne := reference.NewSelf(objectOne)
 	RefTwo := reference.NewSelf(objectTwo)
 
-	rl := NewRequestList()
+	rl := newRequestList()
 	require.Equal(t, 0, rl.Count())
 	require.Equal(t, 0, rl.CountFinish())
 	require.Equal(t, 0, rl.CountActive())
@@ -130,7 +130,7 @@ func TestPendingList_Finish(t *testing.T) {
 	objectTwo := gen.UniqueLocalRefWithPulse(nextPulseNumber)
 	RefTwo := reference.NewSelf(objectTwo)
 
-	rl := NewRequestList()
+	rl := newRequestList()
 
 	require.Equal(t, true, rl.Add(RefOne))
 	require.Equal(t, 1, rl.Count())