**- What I did**
- sync tables behaviour
- hide list method that should not be called outside of table
- added bench for memory usage

before
```
pkg: github.com/insolar/assured-ledger/ledger-core/virtual/object
BenchmarkPendingTable-12    	 3951997	       285 ns/op	     304 B/op	       6 allocs/op
BenchmarkWorkingTable-12    	 4537809	       244 ns/op	     240 B/op	       6 allocs/op
```
after
```
goos: darwin
goarch: amd64
pkg: github.com/insolar/assured-ledger/ledger-core/virtual/object
BenchmarkPendingTable-12    	 7344967	       154 ns/op	     160 B/op	       4 allocs/op
BenchmarkWorkingTable-12    	 5528679	       192 ns/op	     208 B/op	       5 allocs/op
```

**- How I did it**

**- How to verify it**

**- Description for the changelog**